+++ R/fit.R
@@ -143,10 +143,15 @@ fit.MLModelFunction <- function(x, ...) {
 }
 
 
+.fit.TunedModeledRecipe <- function(x, ...) {
+  fit(as(x, "TunedModelRecipe"), model = x@model)
+}
+
+
 eval_fit <- function(data, formula, matrix) {
   use_model_matrix <- if (missing(formula)) TRUE else
     if (missing(matrix)) FALSE else
-      is(terms(data), "DesignTerms")
+      is(terms(data), "ModelDesignTerms")
 
   if (use_model_matrix) {
     envir <- list(