+++ R/resample.R
@@ -55,8 +55,8 @@ resample <- function(x, ...) {
 #'
 resample.formula <- function(x, data, model,
                              control = MachineShop::settings("control"), ...) {
-  resample(ModelFrame(x, data, na.rm = FALSE,
-                      strata = strata(response(x, data))), model, control)
+  mf <- ModelFrame(x, data, na.rm = FALSE, strata = strata(response(x, data)))
+  resample(mf, model, control, ...)
 }
 
 
@@ -64,7 +64,8 @@ resample.formula <- function(x, data, model,
 #'
 resample.matrix <- function(x, y, model,
                             control = MachineShop::settings("control"), ...) {
-  resample(ModelFrame(x, y, na.rm = FALSE, strata = strata(y)), model, control)
+  mf <- ModelFrame(x, y, na.rm = FALSE, strata = strata(y))
+  resample(mf, model, control, ...)
 }
 
 
@@ -80,7 +81,7 @@ resample.ModelFrame <- function(x, model,
                                 control = MachineShop::settings("control"),
                                 ...) {
   if (missing(model)) model <- NullModel
-  .resample(getMLObject(control, "MLControl"), x, model)
+  .resample(getMLObject(control, "MLControl"), x, model, ...)
 }
 
 
@@ -94,7 +95,7 @@ resample.ModelFrame <- function(x, model,
 resample.recipe <- function(x, model,
                             control = MachineShop::settings("control"), ...) {
   if (missing(model)) model <- NullModel
-  .resample(getMLObject(control, "MLControl"), ModelRecipe(x), model)
+  .resample(getMLObject(control, "MLControl"), ModelRecipe(x), model, ...)
 }
 
 
@@ -141,7 +142,8 @@ Resamples.list <- function(object, ...) {
 }
 
 
-.resample.MLBootstrapControl <- function(object, x, model, ...) {
+.resample.MLBootstrapControl <- function(object, x, model, progress_index = 0,
+                                         ...) {
   presets <- MachineShop::settings()
   strata <- strata_var(x)
   set.seed(object@seed)
@@ -155,8 +157,21 @@ Resamples.list <- function(object, ...) {
     train_pred <- subsample(x, x, model, object)$Predicted
   }
 
+  snow_opts <- list()
+  if (MachineShop::settings("progress.resample") && getDoParRegistered()) {
+    total <- if (getDoParName() == "doSNOW") {
+      snow_opts$progress <- function(n) pb$tick()
+      length(splits)
+    } else 0
+    pb <- new_progress_bar(total, input = x, model = model,
+                           index = progress_index)
+    on.exit(pb$terminate())
+  }
+
   foreach(i = seq(splits),
-          .packages = MachineShop::settings("require")) %dopar% {
+          .packages = MachineShop::settings("require"),
+          .verbose = MachineShop::settings("verbose.resample"),
+          .options.snow = snow_opts) %dopar% {
     MachineShop::settings(presets)
     set.seed(seeds[i])
     train <- analysis(splits[[i]], x)
@@ -176,7 +191,8 @@ Resamples.list <- function(object, ...) {
 }
 
 
-.resample.MLCrossValidationControl <- function(object, x, model, ...) {
+.resample.MLCrossValidationControl <- function(object, x, model,
+                                               progress_index = 0, ...) {
   presets <- MachineShop::settings()
   strata <- strata_var(x)
   set.seed(object@seed)
@@ -188,8 +204,21 @@ Resamples.list <- function(object, ...) {
 
   is_optimism_control <- is(object, "MLCVOptimismControl")
 
+  snow_opts <- list()
+  if (MachineShop::settings("progress.resample") && getDoParRegistered()) {
+    total <- if (getDoParName() == "doSNOW") {
+      snow_opts$progress <- function(n) pb$tick()
+      length(splits)
+    } else 0
+    pb <- new_progress_bar(total, input = x, model = model,
+                           index = progress_index)
+    on.exit(pb$terminate())
+  }
+
   df_list <- foreach(i = seq(splits),
-                     .packages = MachineShop::settings("require")) %dopar% {
+                     .packages = MachineShop::settings("require"),
+                     .verbose = MachineShop::settings("verbose.resample"),
+                     .options.snow = snow_opts) %dopar% {
     MachineShop::settings(presets)
     set.seed(seeds[i])
     train <- analysis(splits[[i]], x)
@@ -219,7 +248,7 @@ Resamples.list <- function(object, ...) {
 }
 
 
-.resample.MLOOBControl <- function(object, x, model, ...) {
+.resample.MLOOBControl <- function(object, x, model, progress_index = 0, ...) {
   presets <- MachineShop::settings()
   strata <- strata_var(x)
   set.seed(object@seed)
@@ -227,8 +256,22 @@ Resamples.list <- function(object, ...) {
                        times = object@samples,
                        strata = strata)$splits
   seeds <- sample.int(.Machine$integer.max, length(splits))
+
+  snow_opts <- list()
+  if (MachineShop::settings("progress.resample") && getDoParRegistered()) {
+    total <- if (getDoParName() == "doSNOW") {
+      snow_opts$progress <- function(n) pb$tick()
+      length(splits)
+    } else 0
+    pb <- new_progress_bar(total, input = x, model = model,
+                           index = progress_index)
+    on.exit(pb$terminate())
+  }
+
   foreach(i = seq(splits),
-          .packages = MachineShop::settings("require")) %dopar% {
+          .packages = MachineShop::settings("require"),
+          .verbose = MachineShop::settings("verbose.resample"),
+          .options.snow = snow_opts) %dopar% {
     MachineShop::settings(presets)
     set.seed(seeds[i])
     train <- analysis(splits[[i]], x)
@@ -283,7 +326,7 @@ assessment.ModelFrame <- function(x, object, ...) {
 }
 
 assessment.ModelRecipe <- function(x, object, ...) {
-  rsample::assessment(x)
+  recipe(object, rsample::assessment(x))
 }
 
 
@@ -334,24 +377,29 @@ subsample <- function(train, test, model, control, id = 1) {
 }
 
 
-resample_selection <- function(x, transform, params, ...) {
+resample_selection <- function(x, transform, params, ..., class) {
 
   metrics <- params$metrics
   stat <- fget(params$stat)
 
   perf_list <- list()
   perf_stats <- numeric()
-  for (name in names(x)) {
+  err_msgs <- character()
+  i <- structure(0, max = length(x), names = class)
+  while (i < attr(i, "max")) {
+    i <- i + 1
+    name <- names(x)[i]
+
     res <- try(
-      resample(transform(x[[name]]), ..., control = params$control),
+      resample(transform(x[[name]]), ..., control = params$control,
+               progress_index = i),
       silent = TRUE
     )
 
     if (is(res, "try-error")) {
-      warn("resampling failed for ", name, " with error:\n",
-           attr(res, "condition")$message)
       perf_list[[name]] <- NA
       perf_stats[name] <- NA
+      err_msgs[name] <- attr(res, "condition")$message
       next
     }
 
@@ -372,9 +420,11 @@ resample_selection <- function(x, transform, params, ...) {
   }
 
   failed <- is.na(perf_list)
+  err_msgs <- paste(names(err_msgs), err_msgs, collapse = "\n")
   if (all(failed)) {
-    stop("resampling failed for all models", call. = FALSE)
+    stop("resampling failed for all models\n", err_msgs, call. = FALSE)
   } else if (any(failed)) {
+    warn("resampling failed for some models\n", err_msgs)
     perf[] <- NA
     perf_list[failed] <- list(perf)
   }